+++ app/Auth/Infrastructure/Models/EloquentAdmin.php
@@ -8,22 +8,26 @@ use Filament\Models\Contracts\FilamentUser;
 use Filament\Panel;
 use Illuminate\Foundation\Auth\User as Authenticatable;
 use Illuminate\Notifications\Notifiable;
+use Spatie\Permission\Traits\HasRoles;
 
 class EloquentAdmin extends Authenticatable implements FilamentUser
 {
+    use HasRoles;
     use Notifiable;
 
-    protected $table    = 'admins';
+    protected $table             = 'admins';
+
+    protected string $guard_name = 'admin';
 
     /** @var list<string> */
-    protected $fillable = [
+    protected $fillable          = [
         'name',
         'email',
         'password',
     ];
 
     /** @var list<string> */
-    protected $hidden   = [
+    protected $hidden            = [
         'password',
         'remember_token',
     ];